Normalize types before applying subtyping rules.
This helps recognise distributivity of joins over meets, such as:

  meet<A, join<B, C>> <: join<meet<A, B>, meet<A, C>>

While distributivity of joins over meets alone is quite niche, and
many languages get away without supporting it, distributivity over
viewpoint types (which will be added soon) is a lot more critical.
This PR lays the ground work for that.