Make reachability more standalone.
Reachability was using Generator in order to allocate Label and
Descriptor values upfront. This makes it impossible to reuse the
reachability code for other purposes.

We remove that dependency and don't mix bytecode generation with
reachability. Label and Descriptors are allocated lazily during
bytecode generation, using a new ProgramTable struct.

Additionally, we use this ProgramTable to hold labels to when
closures.